requirements.txt
streamlit==1.11.1
torch==1.13.1
opencv-python-headless==4.5.5.64
pandas==1.3.5
Pillow==8.4.0
numpy==1.7.2
python>=3.7,<3.11
